Move finality test helpers to the test-utils package
Use test-utils test-helpers
